compute: mz_compute_hydration_statuses view
This commit changes `mz_compute_hydration_statuses` from a
storage-managed collection to a view that is defined in terms of
`mz_compute_hydration_times`, and also adds the hydration time as an
additional column. This allows us to rip out some of the hydration
status tracking from the compute controller.