storage: round probe timestamps to the probe interval
This commit ensures that probe timestamps, and therefore timestamps used
for minting new bindings, are rounded down to the probe interval
(typically 1s). This is to reduce the amount of distinct timestamps
flowing through a dataflow joining multiple sources. Each distinct
timestamp induces some amount of overhead, so forcing the timestamps of
individual sources to the same values is more efficient.